+++ src/layouts/MainLayout.jsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { makeStyles } from '@material-ui/core/styles'
 import Sidebar from '../components/Sidebar/Sidebar'
 import Header from '../components/Header/Header'
@@ -11,13 +11,15 @@ const useStyles = makeStyles({
 })
 
 const MainLayout = ({ children }) => {
+	const [sidebar, setSidebar] = useState(false)
+
 	const classes = useStyles()
 
 	return (
 		<div className="Main">
-			{/* <Sidebar /> */}
+			<Sidebar open={sidebar} setOpen={setSidebar} />
 			<div className={classes.container}>
-				<Header />
+				<Header setSidebar={setSidebar} />
 				<main>{children}</main>
 				<Footer />
 			</div>
